requirements.txt
python-dotenv
werkzeug
flask
cryptography